+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Yii Files Change Log
 
-## 1.0.3 under development
+## 1.1.0 under development
 
 - Bug #58: Add missed return value and type for callback of `set_error_handler()` function (vjik)
-- Feature #59: Add new `copyFile` method and `beforeCopy`, `afterCopy` callbacks
+- New #59: Add method `FileHelper::copyFile()` and for method `FileHelper::copyDirectory()` add callbacks
+  `beforeCopy`, `afterCopy`(Gerych1984)
 
 ## 1.0.2 January 11, 2022